pyproject.toml
[project]
name = "sistema-reservas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bcrypt>=5.0.0",
    "django>=6.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "markdown>=3.10",
    "psycopg>=3.3.2",
    "psycopg-binary>=3.3.2",
    "python-dotenv>=1.2.1",
]
